Large refactoring of ocean state ensemble classes, and appropriate changes in notebooks
